mozilla/browser/fuel/test/test_Browser.html

<html>
<head>
  <title>Testing Browser</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"></link>
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

test_Browser();

function test_Browser() {
  netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");

  var browsers = Application.browsers;
  ok(browsers, "Check access to browser windows");
  ok(browsers.length, "There should be at least one browser window open");
  
  var activeBrowser = Application.activeBrowser;
  var moz = activeBrowser.open("http://mozilla.org/");
  var goog = activeBrowser.open("http://google.com/");
  goog.focus();

  is(activeBrowser.tabs.length, 3, "Checking length of 'Browser.tabs' after opening 2 additional tabs");
  is(activeBrowser.activeTab.index, goog.index, "Checking 'Browser.activeTab' after setting focus");
  
  // need to wait for the url's to be refreshed during the load
  function checkURL() {
    is(moz.url, "http://mozilla.org/", "Checking 'BrowserTab.url' after opening");  
    is(goog.url, "http://google.com/", "Checking 'BrowserTab.url' after opening");  
  }
  setTimeout(checkURL, 50);
}
</script>
</pre>
</body>
</html>
